+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Guard STAC metadata parsing against invalid "item_assets" usage ([#853](https://github.com/Open-EO/openeo-python-client/issues/853))
+- `openeo.testing.results`: only do `pixel_tolerance` check on numerical data ([#864](https://github.com/Open-EO/openeo-python-client/issues/864))
 
 
 ## [0.47.0] - 2025-12-17